+++ EmploymentWork/Common/TrapManager.cpp
@@ -323,7 +323,7 @@ void TrapManager::Draw()
 	DrawRotaGraph(static_cast<int>(kTrapPointBgDrawPos.x) - kTrapPointIconOffsetX, static_cast<int>(kTrapPointBgDrawPos.y), 0.66f, 0.0f, m_iconHandle, true);
 
 	auto drawPos = kTrapPointBgDrawPos + kTrapPointOffsetPos;
-	FontManager::GetInstance().DrawBottomRightAndQuakeText(drawPos.x, drawPos.y, std::to_string(m_trapPoint), 0x9effff, 32, m_isTextShake, m_textShakeFrame);
+	FontManager::GetInstance().DrawBottomRightAndQuakeText(static_cast<int>(drawPos.x), static_cast<int>(drawPos.y), std::to_string(m_trapPoint), 0x9effff, 32, m_isTextShake, m_textShakeFrame);
 	//DrawFormatString(76, 720 - 16 * 4, 0xffffff, "%d", m_trapPoint);
 }
